CB-9954	and CB-9953: back-end cloudbreak change to set up separate bucket and role for freeipa auto-backup data
Backups for FreeIPA need to be able to have a user-defined cloud storage location that has a separate policy to be able to be defined for that location.

It should be similar as the top-level telemetry and log storage location that is usable by all SDX and Workloads including both VM and K8s based workloads.