CB-10095 Subnet selection for public gateway endpoint
Adds new logic to create a public load balancer to act as a public gateway to
the Knox endpoints in an otherwise private system. Logic will be triggered if
the "publicEndpointAccessGateway" flag in the environment request is set to
ENABLED, and the tenant has the new CDP_PUBLIC_ENDPOINT_ACCESS_GATEWAY
entitlement.

A list of subnets to use for the public LB can optionally be provided as part
of the environment creation request. If provided, any private subnets in the
list will be discarded, and a public subnet in the same AZ as the selected stack
subnet will be used. If the endpoint subnets aren't explicitly provided, the
subnets provided for environment creation are used instead. If public subnets in
the appropriate availability zones are not provided, creation will fail
immediately.

Tested with unit tests and by running CB locally with hardcoded API fields to
verify cloud resources were created and configured appropriately.